+++ doorstop/core/item.py
@@ -6,12 +6,13 @@ import logging
 
 from doorstop.core import base
 from doorstop.core.base import auto_load, auto_save, BaseFileObject
+from doorstop.core.base import BaseValidatable
 from doorstop import common
 from doorstop.common import DoorstopError, DoorstopWarning, DoorstopInfo
 from doorstop import settings
 
 
-class Item(BaseFileObject):  # pylint: disable=R0904
+class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0904
 
     """Represents an item file with linkable text."""
 
@@ -66,7 +67,7 @@ class Item(BaseFileObject):  # pylint: disable=R0904
         if common.VERBOSITY < common.STR_VERBOSITY:
             return self.id
         else:
-            return self.id_relpath
+            return "{} ({})".format(self.id, self.relpath)
 
     def __eq__(self, other):
         return isinstance(other, self.__class__) and self.path == other.path
@@ -178,12 +179,6 @@ class Item(BaseFileObject):  # pylint: disable=R0904
         """Get the item's relative path string."""
         relpath = os.path.relpath(self.path, self.root)
         return "@{}{}".format(os.sep, relpath)
-
-    # TODO: think of a better name for this property
-    @property
-    def id_relpath(self):
-        """Get the item's ID + relative path string."""
-        return "{} ({})".format(self.id, self.relpath)
 
     @property
     def prefix(self):
@@ -348,44 +343,30 @@ class Item(BaseFileObject):  # pylint: disable=R0904
 
     @auto_save
     @auto_load
-    def add_link(self, item):
-        """Add a new link to another item ID."""
-        self._data['links'].add(item)
+    def link(self, identifier):
+        """Add a new link to another item ID.
+
+        @param identifier: item's ID (or item)
+
+        """
+        identifier = get_id(identifier)
+        self._data['links'].add(identifier)
 
     @auto_save
     @auto_load
-    def remove_link(self, item):
-        """Remove an existing link by item ID."""
-        try:
-            self._data['links'].remove(item)
-        except KeyError:
-            logging.warning("link to {0} does not exist".format(item))
+    def unlink(self, identifier):
+        """Remove an existing link by item ID.
 
-    def valid(self, document=None, tree=None):
-        """Check the item for validity.
-
-        @param document: Document containing the item
-        @param tree: Tree containing the item
-
-        @return: indication that the item is valid
+        @param identifier: item's ID (or item)
 
         """
-        # TODO: refactor: this could be common code with Item/Document/Tree
-        valid = True
-        # Display all issues
-        for issue in self.get_issues(document=document, tree=tree):
-            if isinstance(issue, DoorstopInfo):
-                logging.info(issue)
-            elif isinstance(issue, DoorstopWarning):
-                logging.warning(issue)
-            else:
-                assert isinstance(issue, DoorstopError)
-                logging.error(issue)
-                valid = False
-        # Return the result
-        return valid
+        identifier = get_id(identifier)
+        try:
+            self._data['links'].remove(identifier)
+        except KeyError:
+            logging.warning("link to {0} does not exist".format(identifier))
 
-    def get_issues(self, document=None, tree=None):
+    def get_issues(self, document=None, tree=None, **_):
         """Yield all the item's issues.
 
         @param document: Document containing the item (document-level issues)
@@ -428,11 +409,6 @@ class Item(BaseFileObject):  # pylint: disable=R0904
         # Reformat the file
         if settings.REFORMAT:
             self.save()
-
-    @property
-    def issues(self):
-        """Get a list of the item's issues."""
-        return list(self.get_issues())
 
     def _get_issues_document(self, document):
         """Yield all the item's issues against its document."""
@@ -583,6 +559,11 @@ class Item(BaseFileObject):  # pylint: disable=R0904
 
 
 # attribute formatters #######################################################
+
+def get_id(value):
+    """Get an ID from an item or string."""
+    return str(value).split(' ')[0]
+
 
 def split_id(text):
     """Split an item's ID into a prefix and number.