+++ doorstop/core/tree.py
@@ -3,17 +3,40 @@
 import os
 import sys
 import shutil
+import functools
+from itertools import chain
 import subprocess
 import logging
-from itertools import chain
 
-from doorstop.core.base import clear_document_cache, clear_item_cache
-from doorstop.common import DoorstopError, DoorstopWarning, DoorstopInfo
-from doorstop.core.document import Document
+from doorstop.core.base import BaseValidatable
+from doorstop.common import DoorstopError, DoorstopWarning
+from doorstop.core.document import Document, get_prefix
 from doorstop.core import vcs
 
 
-class Tree(object):  # pylint: disable=R0902
+def clear_document_cache(func):
+    """Decorator for methods that should clear the document cache."""
+    @functools.wraps(func)
+    def wrapped(self, *args, **kwargs):
+        """Wrapped method to clear document cache after execution."""
+        result = func(self, *args, **kwargs)
+        self._document_cache.clear()  # pylint: disable=W0212
+        return result
+    return wrapped
+
+
+def clear_item_cache(func):
+    """Decorator for methods that should clear the item cache."""
+    @functools.wraps(func)
+    def wrapped(self, *args, **kwargs):
+        """Wrapped method to clear item cache after execution."""
+        result = func(self, *args, **kwargs)
+        self._item_cache.clear()  # pylint: disable=W0212
+        return result
+    return wrapped
+
+
+class Tree(BaseValidatable):  # pylint: disable=R0902
 
     """A bidirectional tree structure to store the hierarchy of documents.
 
@@ -164,11 +187,11 @@ class Tree(object):  # pylint: disable=R0902
 
     @clear_document_cache
     @clear_item_cache
-    def new(self, path, prefix, sep=None, digits=None, parent=None):  # pylint: disable=R0913
+    def new_document(self, path, prefix, sep=None, digits=None, parent=None):  # pylint: disable=R0913
         """Create a new document and add it to the tree.
 
         @param path: directory path for the new document
-        @param prefix: document's prefix
+        @param prefix: document's prefix (or document)
         @param sep: separator between prefix and numbers
         @param digits: number of digits for the document's numbers
         @param parent: parent document's prefix
@@ -178,6 +201,7 @@ class Tree(object):  # pylint: disable=R0902
         @raise DoorstopError: if the document cannot be created
 
         """
+        prefix = get_prefix(prefix)
         document = Document.new(path, self.root, prefix,
                                 sep=sep, digits=digits,
                                 parent=parent)
@@ -192,10 +216,10 @@ class Tree(object):  # pylint: disable=R0902
         return document
 
     @clear_item_cache
-    def add(self, prefix, level=None):
+    def add_item(self, prefix, level=None):
         """Add a new item to an existing document by prefix.
 
-        @param prefix: document's prefix
+        @param prefix: document's prefix (or document)
         @param level: desired item level
 
         @return: newly created Item
@@ -203,16 +227,17 @@ class Tree(object):  # pylint: disable=R0902
         @raise DoorstopError: if the item cannot be created
 
         """
+        prefix = get_prefix(prefix)
         document = self.find_document(prefix)
         self.vcs.lock(document.config)  # prevents duplicate item IDs
-        item = document.add(level=level)
+        item = document.add_item(level=level)
         return item
 
     @clear_item_cache
-    def remove(self, identifier):
+    def remove_item(self, identifier):
         """Remove an item from a document by ID.
 
-        @param identifier: item's ID
+        @param identifier: item's ID (or item)
 
         @return: removed Item
 
@@ -225,16 +250,16 @@ class Tree(object):  # pylint: disable=R0902
             except DoorstopError:
                 pass  # item not found in that document
             else:
-                item = document.remove(identifier)
+                item = document.remove_item(identifier)
                 return item
 
         raise DoorstopError("no matching ID: {}".format(identifier))
 
-    def link(self, cid, pid):
+    def link_items(self, cid, pid):
         """Add a new link between two items by IDs.
 
-        @param cid: child item's ID
-        @param pid: parent item's ID
+        @param cid: child item's ID (or child item)
+        @param pid: parent item's ID (or parent item)
 
         @return: child Item, parent Item
 
@@ -247,14 +272,14 @@ class Tree(object):  # pylint: disable=R0902
         # Find parent item
         parent = self.find_item(pid, _kind='parent')
         # Add link
-        child.add_link(parent.id)
+        child.link(parent.id)
         return child, parent
 
-    def unlink(self, cid, pid):
+    def unlink_items(self, cid, pid):
         """Remove a link between two items by IDs.
 
-        @param cid: child item's ID
-        @param pid: parent item's ID
+        @param cid: child item's ID (or child item)
+        @param pid: parent item's ID (or parent item)
 
         @return: child Item, parent Item
 
@@ -267,15 +292,15 @@ class Tree(object):  # pylint: disable=R0902
         # Find parent item
         parent = self.find_item(pid, _kind='parent')
         # Remove link
-        child.remove_link(parent.id)
+        child.unlink(parent.id)
         return child, parent
 
-    def edit(self, identifier, tool=None, launch=False):
+    def edit_item(self, identifier, tool=None, launch=False):
         """Open an item for editing by ID.
 
-        @param identifier: ID of item to edit
+        @param identifier: item's ID (or item)
         @param tool: alternative text editor to open the item
-        @param launch: open the default text editor
+        @param launch: open the text editor
 
         @raise DoorstopError: if the item cannot be found
 
@@ -296,13 +321,14 @@ class Tree(object):  # pylint: disable=R0902
     def find_document(self, prefix):
         """Get a document by its prefix.
 
-        @param prefix: document's prefix
+        @param prefix: document's prefix (or document)
 
         @return: matching Document
 
         @raise DoorstopError: if the document cannot be found
 
         """
+        prefix = get_prefix(prefix)
         logging.debug("looking for document '{}'...".format(prefix))
         try:
             document = self._document_cache[prefix]
@@ -325,7 +351,7 @@ class Tree(object):  # pylint: disable=R0902
     def find_item(self, identifier, _kind=''):
         """Get an item by its ID.
 
-        @param identifier: item ID
+        @param identifier: item's ID (or item)
 
         @return: matching Item
 
@@ -356,32 +382,7 @@ class Tree(object):  # pylint: disable=R0902
 
         raise DoorstopError("no matching{} ID: {}".format(_kind, identifier))
 
-    def valid(self, document_hook=None, item_hook=None):
-        """Check the tree (and its documents) for validity.
-
-        @param document_hook: function to call for custom document validation
-        @param item_hook: function to call for custom item validation
-
-        @return: indication that the tree is valid
-
-        """
-        valid = True
-        logging.info("checking tree...")
-        # Display all issues
-        for issue in self.get_issues(document_hook=document_hook,
-                                     item_hook=item_hook):
-            if isinstance(issue, DoorstopInfo):
-                logging.info(issue)
-            elif isinstance(issue, DoorstopWarning):
-                logging.warning(issue)
-            else:
-                assert isinstance(issue, DoorstopError)
-                logging.error(issue)
-                valid = False
-        # Return the result
-        return valid
-
-    def get_issues(self, document_hook=None, item_hook=None):
+    def get_issues(self, document_hook=None, item_hook=None, **_):
         """Yield all the tree's issues.
 
         @param document_hook: function to call for custom document validation
@@ -403,11 +404,6 @@ class Tree(object):  # pylint: disable=R0902
                 # Prepend the document's prefix to yielded exceptions
                 if isinstance(issue, Exception):
                     yield type(issue)("{}: {}".format(document.prefix, issue))
-
-    @property
-    def issues(self):
-        """Get a list of the tree's issues."""
-        return list(self.get_issues())
 
     @clear_document_cache
     @clear_item_cache